requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
asyncpg==0.29.0
python-dotenv==1.0.1
apscheduler==3.10.4
openpyxl==3.1.2
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
psycopg2-binary==2.9.11
